requirements.txt
streamlit
pandas
requests
openai
watchdog
mplsoccer
matplotlib
pycountry
faiss-cpu
numpy